requirements.txt
pillow
requests
flask
opencv-python
pytesseract
numpy
pytube
flask-socketio
yt_dlp
